+++ swov1/RELEASES.md
@@ -95,3 +95,13 @@ Based on:
 - [typescript v0.8.1] swov1
 ### Releases
 - [NPM v0.8.1] https://www.npmjs.com/package/@solarwinds/swo-sdk-typescript/v/0.8.1 - swov1
+
+## 2026-02-24 00:06:21
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.725.0 (2.836.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.9.0] swov1
+### Releases
+- [NPM v0.9.0] https://www.npmjs.com/package/@solarwinds/swo-sdk-typescript/v/0.9.0 - swov1